Rename layout inflater factory
We don't have anything specifically to do with Dagger so it shouldn't be in the name.
