+++ inflation-inject-sample/src/main/java/com/example/MainActivity.java
@@ -2,15 +2,14 @@ package com.example;
 
 import android.app.Activity;
 import android.os.Bundle;
-import android.view.LayoutInflater;
-import com.squareup.inject.inflation.DaggerLayoutInflaterFactory;
+import com.squareup.inject.inflation.InflationInjectFactory;
 import dagger.Component;
 
 public final class MainActivity extends Activity {
   @Override protected void onCreate(Bundle savedInstanceState) {
     super.onCreate(savedInstanceState);
 
-    DaggerLayoutInflaterFactory factory = DaggerMainActivity_MainComponent.create().factory();
+    InflationInjectFactory factory = DaggerMainActivity_MainComponent.create().factory();
     getLayoutInflater().setFactory(factory);
 
     setContentView(R.layout.custom_view);
@@ -18,6 +17,6 @@ public final class MainActivity extends Activity {
 
   @Component(modules = ViewModule.class)
   interface MainComponent {
-    DaggerLayoutInflaterFactory factory();
+    InflationInjectFactory factory();
   }
 }
